Refactoring of session instance handling
Introduce new class SessioWrapper that handles and loads the session.
It will also pass all relevant information to the session and is used to
encapsuate a session instance to be used in the SessionManager.
SessionManager will resolve and take care of dependencies and parallel
builds, etc. in the future. Also renamed args replace methods of Session
to be more public