Add __getattribute__ method for Session class
Overwrite object __getattribute__ method to be able to replace values
for src_patch, build_path, etc. on the fly. Also add a feature that
allows to specifiy dependecies in an parent class which will be
recognized in the child too. This is usefull for CMakeSession for
example because all Sessions based on that parent Session should depend
on cmake too.