Set session.name if it is not set
Use the first part of the module name to set the session name
automatically. Session name was introduced to create the name for the
exprected directories and to be able to change the build and source dir
within a session. Therefore it had to be set by the user. Now this isn't
necessary anymore.