Introduce current dir like in mac os frameworks
current dir in a destroot of a session now symlinks to the current used
version of a session. Also in the activate phase already existing files
in the prefix dir will be deleted. In future jam must check if the
deleted file is owned by the session that is being activated.