Add phase classes to reflect a session phase
Phases are necessary to store the current "status" of a session e.g. if
it is already built. If a session is already built jam should not run
previous and current phases again. E.g if a session is already build and
the build command is called, jam should not run anything and display
only a message for the user.