Add Sequence classes
Sequence classes are intended to call only session methods which belong
to a phase. E.g. for install command several methods in a session must
be called. Therefore a sequence object should contain all these methods
which are called after each other. Because it isn't necessary to call
e.g. patch again if the current session already patched the sources the
sequence of called methods depends on the current status of a session.
This status is reflected by the phase.