Add shadow dict for arg variables
Some variables in a Session class like args can contain lists of string
including jam variables that must be replaced with a value of another
variable e.g. %(src_path)/mypath. Before jam always returned a new list
with replaced variables. Now it's always returning the same list.
This doesn't solve the problem of appending values to the list. Appended
values are still not available in a second call of the list.