Use Sequences to call SessionWrapper methods
SessionManager now defines Sequences of commands/methods which will be
run if e.g. install is called. Sequences care about the current status
of a Session and only run commands which aren't already executed
priviously. E.g. if build was already called and now a user wants to
install a session only destroot and activate are called.

Also depenencies of a session are installed if configure (or "higher"
methods like build or install) is called.