Remove patch_dir from sessionwrapper and session classes
patch_dir was introduced to point to a directory where the patches are
downloaded. Because patches can't be downloaded anymore this variable is
obsolete. A new varibale patch_path is introduced. The patch_path
variable can be set in the session like src_path. It defaults to
"%(session_path)s/patches". This means the subdir patches of the build
session.